# openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW BFF API
  description: |
    Backend for Frontend (BFF) REST API que agrega e simplifica o acesso aos microserviços do sistema ConstrSW.
    
    Esta API atua como uma camada intermediária entre o frontend e os microserviços, fornecendo:
    - Agregação de dados de múltiplos serviços
    - Simplificação de chamadas complexas
    - Redução do número de requisições necessárias
    - Padronização de respostas
    
    ## Autenticação
    Todos os endpoints (exceto `/auth/login` e `/auth/refresh`) requerem autenticação via Bearer Token (JWT).
    O token deve ser obtido através do endpoint `/auth/login` e incluído no header `Authorization: Bearer <token>`.
    
    ## Microserviços Agregados
    - OAuth (Autenticação e Autorização)
    - Employees (Funcionários)
    - Classes (Turmas)
    - Courses (Disciplinas)
    - Lessons (Aulas)
    - Professors (Professores)
    - Reservations (Reservas)
    - Resources (Recursos)
    - Rooms (Salas)
    - Students (Estudantes)
  version: 1.0.0
  contact:
    name: ConstrSW Team
    email: support@constrsw.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desenvolvimento local
  - url: http://api:3000/api/v1
    description: Servidor interno (Docker)
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Listar usuários
      description: Retorna uma lista de usuários cadastrados
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - name: enabled
          in: query
          description: Filtrar por status (true/false)
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Users
      summary: Criar usuário
      description: Cria um novo usuário no sistema
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /users/{userId}/roles:
    get:
      tags:
        - Users
      summary: Listar roles de um usuário
      operationId: getUserRoles
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Lista de roles do usuário
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Users
      summary: Atribuir roles a um usuário
      operationId: assignRolesToUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roleIds
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Roles atribuídos com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Users
      summary: Remover roles de um usuário
      operationId: removeRolesFromUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - roleIds
              properties:
                roleIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Roles removidos com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}:
    patch:
      tags:
        - Users
      summary: Atualizar senha de um usuário
      description: Atualiza a senha de um usuário
      operationId: updateUserPassword
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - password
              properties:
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Senha atualizada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /students:
    get:
      tags:
        - Students
      summary: Listar estudantes
      description: Retorna uma lista de estudantes com filtros opcionais
      operationId: listStudents
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: enrollment
          in: query
          schema:
            type: string
        - name: email
          in: query
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Lista de estudantes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Students
      summary: Criar estudante
      operationId: createStudent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentCreateRequest'
      responses:
        '201':
          description: Estudante criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /students/{studentId}:
    get:
      tags:
        - Students
      summary: Buscar estudante por ID
      operationId: getStudentById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '200':
          description: Dados do estudante
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Students
      summary: Atualizar estudante (completo)
      operationId: updateStudent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentUpdateRequest'
      responses:
        '204':
          description: Estudante atualizado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Students
      summary: Atualizar estudante (parcial - JSON Patch)
      operationId: patchStudent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                type: object
                properties:
                  op:
                    type: string
                    enum: [add, remove, replace, copy, move, test]
                  path:
                    type: string
                  value:
                    type: string
      responses:
        '200':
          description: Estudante atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Students
      summary: Deletar estudante
      operationId: deleteStudent
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/StudentId'
      responses:
        '204':
          description: Estudante deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /rooms:
    get:
      tags:
        - Rooms
      summary: Listar salas
      description: Retorna uma lista de salas com filtros e paginação
      operationId: listRooms
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: building
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, OCCUPIED, MAINTENANCE, RESERVED]
        - name: minCapacity
          in: query
          schema:
            type: integer
        - name: maxCapacity
          in: query
          schema:
            type: integer
        - name: number
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Lista de salas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoomResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Rooms
      summary: Criar sala
      operationId: createRoom
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Sala criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /rooms/{roomId}:
    get:
      tags:
        - Rooms
      summary: Buscar sala por ID
      operationId: getRoomById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Dados da sala
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Rooms
      summary: Atualizar sala (completo)
      operationId: updateRoom
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Sala atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Rooms
      summary: Atualizar sala (parcial)
      operationId: patchRoom
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Sala atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoomResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Rooms
      summary: Deletar sala
      operationId: deleteRoom
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '204':
          description: Sala deletada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /auth/login
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      description: ID único do usuário
      schema:
        type: string
...
    StudentId:
      name: studentId
      in: path
      required: true
      description: ID único do estudante (UUID)
      schema:
        type: string
        format: uuid
...
    RoomId:
      name: roomId
      in: path
      required: true
      description: ID único da sala
      schema:
        type: string
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Mensagem de erro
        status_code:
          type: integer
          description: Código HTTP do erro
    UserResponse:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserCreateRequest:
      type: object
      required:
        - username
        - email
        - password
        - firstName
        - lastName
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        firstName:
          type: string
        lastName:
          type: string
        enabled:
          type: boolean
          default: true
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    StudentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        enrollment:
          type: string
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    StudentCreateRequest:
      type: object
      required:
        - name
        - enrollment
        - email
      properties:
        name:
          type: string
        enrollment:
          type: string
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
    StudentUpdateRequest:
      type: object
      properties:
        name:
          type: string
        enrollment:
          type: string
        email:
          type: string
          format: email
        birthDate:
          type: string
          format: date
...
    RoomResponse:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        building:
          type: string
        category:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, MAINTENANCE, RESERVED]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoomCreateRequest:
      type: object
      required:
        - number
        - building
        - capacity
      properties:
        number:
          type: string
        building:
          type: string
        category:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, MAINTENANCE, RESERVED]
          default: AVAILABLE
    RoomUpdateRequest:
      type: object
      properties:
        number:
          type: string
        building:
          type: string
        category:
          type: string
        capacity:
          type: integer
        status:
          type: string
          enum: [AVAILABLE, OCCUPIED, MAINTENANCE, RESERVED]
...
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autorizado - Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'